+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 For a full diff see [`0.11.0...master`][0.11.0...master].
 
+### Added
+
+* Added `SchemaValidator::validate()`, which returns a `Result` composing validation error messages ([#268]), by [@localheinz]
+
 ## [`0.11.0`][0.11.0]
 
 For a full diff see [`0.10.1...0.11.0`][0.10.1...0.11.0].
@@ -307,6 +311,7 @@ For a full diff see [`5d8b3e2...0.1.0`][5d8b3e2...0.1.0].
 [#191]: https://github.com/ergebnis/json-normalizer/pull/191
 [#202]: https://github.com/ergebnis/json-normalizer/pull/202
 [#203]: https://github.com/ergebnis/json-normalizer/pull/203
+[#268]: https://github.com/ergebnis/json-normalizer/pull/268
 
 [@BackEndTea]: https://github.com/BackEndTea
 [@ergebnis]: https://github.com/ergebnis